requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
django-cors-headers>=4.3.0
dj-database-url>=2.1.0
django-filter>=23.0,<25.0
